+++ .storybook/preview-head.html
@@ -1,1 +1,3 @@
+<link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.6/dist/web/static/pretendard.css" />
+
 <div id="portal"></div>
