+++ package.json
@@ -24,6 +24,7 @@
     "@storybook/addon-interactions": "^7.0.24",
     "@storybook/addon-links": "^7.0.24",
     "@storybook/addon-mdx-gfm": "^7.0.24",
+    "@storybook/addon-styling": "^1.3.2",
     "@storybook/blocks": "^7.0.24",
     "@storybook/react": "^7.0.24",
     "@storybook/react-vite": "^7.0.24",
