+++ src/styles/theme.js
@@ -3,6 +3,10 @@ const theme = {
     bg: "#ffffff",
     black: "#111111",
   },
+
+  fontFamily: {
+    main: '"Pretendard Variable", Pretendard, -apple-system, BlinkMacSystemFont, system-ui, Roboto, "Helvetica Neue", "Segoe UI", "Apple SD Gothic Neo", "Noto Sans KR", "Malgun Gothic", "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", sans-serif',
+  },
 };
 
 export default theme;
